split current routes
